requirements.txt
fastapi
uvicorn
pandas
numpy
scikit-learn
joblib
pyyaml
requests
